updates component
