+++ components/wallet/index.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+const ConnetWallet = () => {
+  return (
+    <a
+      href="#"
+      className="js-wallet group flex h-10 px-6 items-center justify-center rounded-full border border-transparent font-display bg-accent text-white transition-colors hover:bg-white hover:text-accent hover:border-jacarta-100 dark:border-transparent dark:bg-white/[.15] dark:hover:bg-accent"
+      aria-label="connect"
+    >
+      Connect
+    </a>
+  );
+};
+
+export default ConnetWallet;
