+++ layouts/AppLayout.tsx
@@ -1,16 +1,17 @@
-import { ReactNode, useEffect, useState } from 'react';
-import Header from '../components/header'
+import { ReactNode, useEffect, useState } from "react";
+import Hero from "../components/hero";
+import Header from "../components/header";
 
 interface Props {
   pageTitle?: string | null;
   children: ReactNode;
 }
 
-const AppLayout = ({pageTitle, children}: Props): JSX.Element => {
-
+const AppLayout = ({ pageTitle, children }: Props): JSX.Element => {
   return (
     <div id="app">
       <Header />
+      {/* <Hero /> */}
       {/* <main>{children}</main> */}
     </div>
   );
